Fix bug where count text was not updating
The count and position text were not updating in the ContentView because
the PushupsDetector object was not correctly injected into the
environment. Removing @Observable and treating PushupsDetector instead
as an ObservableObject (not sure how it came to be declared with both)
fixes the issue such that updates are now propagated to the ContentView
while the user is doing pushups.