Update pitch/acc data from PushupsDetector
The pitch and acceleration data was not previously being recorded. While
the MotionManager properties for pitch and acceleration arrays were being
persisted, these properties were never actually mutated (outside of test).
So in order to record the correct values for these properties I moved
them to the PushupsDetector. It's worth noting that these solution is
incomplete-- I needed to use a reference to DataStorage that is NOT an
@EnvironmentObject (because PushupsDetector cannot contain such a
reference), so this causes some issues with the avgPitch, which is WIP.