Access Exercise from CountVC directly, not Environment
This circumvents the issue of having two environment objects of the same
type (Exercise), in part because these don't need to be environment
objects at all. I also condensed both detectors (squat and pushup)
into one and then init the detector with the correct type when the CountViewController is
initialized.