Make ResizingButton title into a Binding
This solves the issue with how to update the text in the
ExerciseCountViewController, but it is a bit clunky. Could probably be
refactored again to extract out the button text altogether, which should
remove the need for a Binding and extra fiddly button title properties
to make bindings out of regular Strings.